feat: do not remove fill attributes for credit card icons
We remove the `fill` attributes of a svg if all elements excluding
`none`, `white`, and `black` filled elements are the same fill color,
and then we color icons afterwards.

For credit cards that fill color in ways like patterns, this might not
work for the same goal. Given that we'll not color credit card icons
programmatically, this skip the processing for credit card icons, and
simply repsect the color from the SVG files.